minor adjustments

+++ ST_engine/ST_message_bus/include/message_bus.hpp
@@ -30,13 +30,13 @@
 class message_bus{
     private:
         friend class message_bus_tests;
-        ska::bytell_hash_map<int, std::vector<subscriber*>> subscribers; //each message enum maps to a list of subscribers for that message
+        ska::bytell_hash_map<uint8_t, std::vector<subscriber*>> subscribers; //each message enum maps to a list of subscribers for that message
     public:
         message_bus();
         ~message_bus();
         void clear();
         void send_msg(message* msg);
-        void subscribe(int msg, subscriber* sub);
+        void subscribe(uint8_t msg, subscriber* sub);
 };
 
 /**
@@ -49,7 +49,7 @@ class message_bus{
  * @param data The data the message carries - created with <b>make_data<>()</b> or is <b>nullptr</b>
  * @return A new message object.
  */
-inline message* make_msg(int name, const std::shared_ptr<void>& data){
+inline message* make_msg(uint8_t name, const std::shared_ptr<void>& data){
     return msg_memory.allocate_message(name, data);
 }
